Move interact to its own file
