+++ src/macOSNotif.js
@@ -19,113 +19,7 @@
 require("./css/macOSNotif.css");  // Ensure the CSS gets included
 require("@babel/polyfill");  // Polyfill for older browsers
 
-class __macOSNotifJSInteract {
-
-    constructor(element) {
-        // Get the actual element (supports selector passing)
-        this.element = typeof element === "string" ? document.querySelector(element) : element;
-
-        this.dragActing = false;
-        this.dragXOrg = null;
-        this.dragXOffset = 0;
-    }
-
-    onDismiss(callback) {
-        // Set the onDismiss action (overwrites this func, so can only be set once)
-        this.onDismiss = callback;
-        return this;
-    }
-
-    scrollMove(evt) {
-        // TODO: make this work, detect X scrolling as a swipe to dismiss
-        if (!evt.deltaX) return;
-        if (evt.deltaX < 0) this.onDismiss();
-    }
-
-    scrollRun() {
-        this.element.addEventListener("wheel", evt => this.scrollMove(evt), true);
-    }
-
-    dragMove(evt) {
-        // Don't run if currently doing drag stuff
-        if (!this.dragActing) return;
-        evt.preventDefault();
-        evt.stopPropagation();
-
-        // Get the position and adjust based on movement
-        let position = this.dragXOffset + this.dragXOrg;
-        if (evt.type === "mousemove") {
-            position -= evt.clientX;
-        } else if (evt.type === "touchmove") {
-            position -= evt.targetTouches[0].clientX;
-        }
-
-        // Only allow dragging to the right of the original notif position
-        if (position > this.dragXOrg) position = this.dragXOrg;
-
-        // Move the element to match mouse drag
-        this.element.style.transition = "unset";
-        this.element.style.right = position + "px";
-    }
-
-    dragRightOffset() {
-        const thisRect = this.element.getBoundingClientRect();
-        const parentRect = this.element.parentElement.getBoundingClientRect();
-        return parentRect.right - thisRect.right;
-    }
-
-    dragStart(evt) {
-        evt.preventDefault();
-        evt.stopPropagation();
-
-        if (evt.type === "mousedown") {
-            this.dragXOffset = evt.clientX;
-        } else if (evt.type === "touchstart") {
-            this.dragXOffset = evt.targetTouches[0].clientX;
-        }
-
-        if (this.dragXOrg === null) this.dragXOrg = this.dragRightOffset();
-        this.dragActing = true;
-    }
-
-    dragStop(evt) {
-        if (!this.dragActing) return;
-        evt.preventDefault();
-        evt.stopPropagation();
-
-        // Reset transition and stop dragging
-        this.element.style.transition = "";
-        this.dragActing = false;
-
-        // Check if we should dismiss
-        const offset = Math.abs(this.dragRightOffset());
-        const threshold = this.element.offsetWidth * 0.2;
-        if (offset >= threshold) {
-            this.onDismiss();
-        } else {
-            this.element.style.right = this.dragXOrg + "px";
-        }
-    }
-
-    dragRun() {
-        this.element.addEventListener("mousedown", evt => this.dragStart(evt), true);
-        this.element.addEventListener("touchstart", evt => this.dragStart(evt), true);
-
-        window.addEventListener("mousemove", evt => this.dragMove(evt), true);
-        window.addEventListener("touchmove", evt => this.dragMove(evt), true);
-
-        window.addEventListener("mouseup", evt => this.dragStop(evt), true);
-        window.addEventListener("touchend", evt => this.dragStop(evt), true);
-    }
-
-    run() {
-        this.dragRun();
-
-        // TODO: this
-        //this.scrollRun();
-    }
-}
-
+const __macOSNotifJSInteract = require("./interact.js");
 const __macOSNotifJSTemplate = require("./html/macOSNotif.html").default.replace(/<!--(?!>)[\S\s]*?-->/g, ""); // Strip HTML comments
 const __macOSNotifJSNotifs = {};
 let __macOSNotifJSFadeThreshold = 6;
